fix various exceptions
- invalid cast in ResolutionEnumToString
- invalid method of instantiating options list in both HandleResolutionElements methods
- unassigned variables in scene

+++ Assets/Code/GameConstants.cs
@@ -169,7 +169,7 @@ namespace SFBuilder
         public static List<string>  ResolutionEnumToString(int rangeStart = 0, int rangeCount = 255)
         {
             return System.Enum.GetValues(typeof(ResolutionSetting))
-                .Cast<int>()
+                .Cast<byte>()
                 .Select(x => x.ToString())
                 .ToList().GetRange(rangeStart, rangeCount);
         }
